+++ src/node/behaviors/__init__.py
@@ -10,10 +10,10 @@ from .cache import VolatileStorageInvalidate  # noqa
 from .common import AsAttrAccess  # noqa
 from .common import ChildFactory  # noqa
 from .common import FixedChildren  # noqa
-from .common import GetattrChildren  # noqa
-from .common import NodeChildValidate  # noqa
 from .common import UnicodeAware  # noqa
 from .common import UUIDAware  # noqa
+from .constraints import MappingConstraints  # noqa
+from .constraints import SequenceConstraints  # noqa
 from .context import BoundContext  # noqa
 from .events import EventAttribute  # noqa
 from .events import Events  # noqa
@@ -59,13 +59,19 @@ deprecated(
     '``Adopt`` has been renamed to ``MappingAdopt``. Please fix your import',
     Adopt='node.behaviors.adopt:MappingAdopt',
 )
-# B/C 2022-02-14 -> node.behaviors.Storage
+# B/C 2022-02-22 -> node.behaviors.NodeChildValidate
 deprecated(
-    '``Storage`` has been renamed to ``MappingStorage``. Please fix your import',
-    Storage='node.behaviors.storage:MappingStorage',
+    '``NodeChildValidate`` has been renamed to ``MappingConstraints``. '
+    'Please fix your import',
+    NodeChildValidate='node.behaviors.constraints:MappingConstraints',
 )
 # B/C 2022-02-14 -> node.behaviors.Nodify
 deprecated(
     '``Nodify`` has been renamed to ``MappingNode``. Please fix your import',
     Nodify='node.behaviors.mapping:MappingNode',
+)
+# B/C 2022-02-14 -> node.behaviors.Storage
+deprecated(
+    '``Storage`` has been renamed to ``MappingStorage``. Please fix your import',
+    Storage='node.behaviors.storage:MappingStorage',
 )