+++ src/node/testing/env.py
@@ -3,7 +3,7 @@ from node.behaviors import MappingAdopt
 from node.behaviors import AsAttrAccess
 from node.behaviors import DefaultInit
 from node.behaviors import MappingNode
-from node.behaviors import NodeChildValidate
+from node.behaviors import MappingConstraints
 from node.behaviors import OdictStorage
 from node.interfaces import INode
 from plumber import plumbing
@@ -21,7 +21,7 @@ class NoNode(object):
 
 
 @plumbing(
-    NodeChildValidate,
+    MappingConstraints,
     DefaultInit,
     MappingAdopt,
     AsAttrAccess,